split fhir exceptions from regular ones
